requirements.txt
slack_bolt
google-cloud-aiplatform
python-dotenv
requests
python-docx
PyPDF2